Introduce single-file upload.
This introduces the previously discussed stanalone cli with the
standard single file upload tooling. It also adds a version of
a streaming upload tool, however it isn't going work as well as
the file upload becase it has to buffer + hash in memory and
can't exploit parallel uploads. This could be improved by streaming
to disk as a better buffer and allocating chunks as we go (but
then why not just use file upload, so not sure if this will live
forever or not).